+++ service/auth/src/main/java/com/webest/auth/presentation/dto/response/JoinResponse.java
@@ -9,6 +9,6 @@ public record JoinResponse(
         String userId,
         String userName,
         String email,
-        Integer addressCode
+        Long addressCode
 ) {
 }
